requirements.txt
apispec==6.10.0
APScheduler==3.11.2
attrs==26.1.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
flasgger==0.9.7.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
flask-smorest==0.47.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
marshmallow==4.3.0
mistune==3.2.0
packaging==26.2
PyJWT==2.12.1
python-dotenv==1.0.1
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
six==1.17.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
webargs==8.7.1
Werkzeug==3.1.6
